+++ checkpoint/orbax/checkpoint/type_handlers.py
@@ -21,7 +21,6 @@ import copy
 import dataclasses
 import functools
 import json
-import math
 import os
 import re
 import sys
@@ -38,6 +37,7 @@ import numpy as np
 from orbax.checkpoint import future
 from orbax.checkpoint import multihost
 from orbax.checkpoint import serialization
+from orbax.checkpoint._src.arrays import subchunking
 from orbax.checkpoint.metadata import sharding as sharding_metadata
 from orbax.checkpoint.metadata import value as value_metadata
 from orbax.checkpoint.path import async_utils
@@ -268,109 +268,6 @@ class RestoreArgs:
   dtype: Optional[jnp.dtype] = None
 
 
-def _find_divisors(size: int):
-  """Fast-ish method for finding divisors of a number."""
-  sqrt_divs = [
-      i for i in range(1, math.ceil(math.sqrt(size + 1))) if size % i == 0
-  ]
-  return sorted(set(sqrt_divs + [size // div for div in sqrt_divs][::-1]))
-
-
-def _choose_chunk_shape(
-    global_shape: Sequence[int],
-    write_shape: Sequence[int],
-    dtype: Union[jnp.dtype, np.dtype],
-    target_byte_size: int,
-) -> Sequence[int]:
-  """Chooses a chunk shape that divides the `write_shape`.
-
-  The chunk shape is chosen such that the resulting byte size is less than
-  or equal to `target_byte_size`, but is otherwise as large as possible.
-
-  This uses a greedy algorithm that attempts to split the largest and sharded
-  dimensions first.
-
-  Args:
-    global_shape: the global shape of the array
-    write_shape: the local shape being written
-    dtype: the dtype of the array
-    target_byte_size: Desired chunk byte size.  Must be >= dtype.itemsize.
-
-  Returns:
-    List of length `len(write_shape)` specifying the chosen chunk shape.
-  """
-  assert len(global_shape) == len(write_shape)
-  if target_byte_size < 1048576:  # 1 MB
-    logging.warning(
-        'Setting the target_byte_size too small could reduce performance.'
-    )
-
-  sharded_dimensions = np.array(global_shape) != np.array(write_shape)
-  dtype_size = dtype.itemsize
-  target_elements = target_byte_size // dtype_size
-
-  rank = len(write_shape)
-
-  # `dim_factors[i]` is the list of divisors of `write_shape[i]`
-  dim_factors = [_find_divisors(size) for size in write_shape]
-
-  # The current chunk shape is:
-  # [dim_factors[i][-1] for i in range(rank)]
-
-  def get_total_elements():
-    """Returns the number of elements in the current chunk shape."""
-    total_elements = 1
-    for i in range(rank):
-      total_elements *= dim_factors[i][-1]
-    return total_elements
-
-  # Reduce the current chunk shape until the desired number of elements is
-  # reached.
-  while get_total_elements() > target_elements:
-    # Greedily reduce the largest dimension.  This is not guaranteed to bring us
-    # the closest to `target_elements`, but is simple to implement and should
-    # work well enough.
-    dim_to_reduce = -1
-    dim_to_reduce_size = 1
-    for i in range(rank):
-      size = dim_factors[i][-1]
-      if sharded_dimensions[i] and size > dim_to_reduce_size:
-        dim_to_reduce_size = size
-        dim_to_reduce = i
-
-    if dim_to_reduce_size > 1:
-      dim_factors[dim_to_reduce].pop()
-    else:
-      # need to start splitting on unsharded dimension
-      sharded_dimensions = np.ones(len(write_shape))
-
-  chosen_shape = [dim_factors[i][-1] for i in range(rank)]
-
-  logging.vlog(
-      1,
-      'global_shape=%s, write_shape=%s, dtype=%s, target_byte_size=%d,'
-      ' chosen_shape=%s',
-      global_shape,
-      write_shape,
-      dtype,
-      target_byte_size,
-      chosen_shape,
-  )
-
-  return chosen_shape
-
-
-def _validate_divisible_shapes(
-    shape1: tuple[int, ...], shape2: tuple[int, ...]
-) -> bool:
-  """Return True if shape2 is a divisor of shape1 otherwise False."""
-  try:
-    return not np.mod(shape1, shape2).any()
-  except ValueError:
-    # eg. imcompatible shape
-    return False
-
-
 # TS functions
 # TODO(b/336658919) refractor TS functions to a separate file
 def _get_json_tspec(
@@ -492,7 +389,7 @@ def _build_ts_zarr_shard_and_chunk_metadata(
   if not use_zarr3:
     # Zarr ver2
     if chunk_byte_size is not None:
-      metadata['chunks'] = _choose_chunk_shape(
+      metadata['chunks'] = subchunking.choose_chunk_shape(
           global_shape, shard_shape, dtype, chunk_byte_size
       )
       logging.info('Chose a chunk shape equal to: %s', str(metadata['chunks']))
@@ -511,7 +408,7 @@ def _build_ts_zarr_shard_and_chunk_metadata(
           raise ValueError(
               f'chunk_byte_size={chunk_byte_size} must be >= {dtype.itemsize}'
           )
-        write_chunk_shape = read_chunk_shape = _choose_chunk_shape(
+        write_chunk_shape = read_chunk_shape = subchunking.choose_chunk_shape(
             global_shape, shard_shape, dtype, chunk_byte_size
         )
       else:
@@ -522,7 +419,9 @@ def _build_ts_zarr_shard_and_chunk_metadata(
 
     # If `write_chunk_shape` is not specified, make it the same as Jax sharding
     if write_chunk_shape:
-      if not _validate_divisible_shapes(shard_shape, write_chunk_shape):
+      if not subchunking.validate_divisible_shapes(
+          shard_shape, write_chunk_shape
+      ):
         raise ValueError(
             f'write_chunk_shape={write_chunk_shape} is not a divisor of'
             f' shard_shape={shard_shape}'
@@ -542,7 +441,9 @@ def _build_ts_zarr_shard_and_chunk_metadata(
     # If `read_chunk_shape` is not specified, make it the same as
     # `write_shape`
     if read_chunk_shape:
-      if not _validate_divisible_shapes(write_shape, read_chunk_shape):
+      if not subchunking.validate_divisible_shapes(
+          write_shape, read_chunk_shape
+      ):
         raise ValueError(
             f'read_chunk_shape={read_chunk_shape} is not a divisor of'
             f' write_chunk_shape={write_shape}'